Extract interface
